refactor(web): move-WithHelpTooltip-to-components
