+++ web/src/pages/Dashboard/JurorInfo/StakingRewards.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { Box as _Box, Button } from "@kleros/ui-components-library";
 import TokenRewards from "./TokenRewards";
-import WithHelpTooltip from "../WithHelpTooltip";
+import WithHelpTooltip from "components/WithHelpTooltip";
 import { EnsureChain } from "components/EnsureChain";
 
 const Container = styled.div`
