+++ web/src/pages/Home/TopJurors/Header/Coherency.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled, { css } from "styled-components";
 import { BREAKPOINT_LANDSCAPE, landscapeStyle } from "styles/landscapeStyle";
 import { useWindowSize } from "react-use";
-import WithHelpTooltip from "pages/Dashboard/WithHelpTooltip";
+import WithHelpTooltip from "components/WithHelpTooltip";
 
 const Container = styled.div`
   display: flex;
