+++ web/src/pages/Home/TopJurors/Header/MobileHeader.tsx
@@ -6,22 +6,21 @@ import HowItWorks from "components/HowItWorks";
 import JurorLevels from "components/Popup/MiniGuides/JurorLevels";
 
 const Container = styled.div`
-display: flex;
-justify-content: space-between;
-width: 100%;
-background-color: ${({ theme }) => theme.lightBlue};
-padding: 24px;
-border 1px solid ${({ theme }) => theme.stroke};
-border-top-left-radius: 3px;
-border-top-right-radius: 3px;
-border-bottom: none;
-flex-wrap: wrap;
-${landscapeStyle(
-  () =>
-    css`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  background-color: ${({ theme }) => theme.lightBlue};
+  padding: 24px;
+  border 1px solid ${({ theme }) => theme.stroke};
+  border-top-left-radius: 3px;
+  border-top-right-radius: 3px;
+  border-bottom: none;
+  flex-wrap: wrap;
+  ${landscapeStyle(
+    () => css`
       display: none;
     `
-)}
+  )}
 `;
 
 const StyledLabel = styled.label`
